+++ plugin-build/src/test/kotlin/io/sentry/android/gradle/SentryPluginVariantTest.kt
@@ -207,7 +207,8 @@ class SentryPluginVariantTest(
             arrayOf("7.0.3", "7.0.2"),
             arrayOf("7.0.3", "7.1.1"),
             arrayOf("7.0.3", "7.2"),
-            arrayOf("7.1.0-alpha13", "7.2")
+            arrayOf("7.1.0-beta01", "7.2"),
+            arrayOf("7.2.0-alpha01", "7.2")
         )
 
         private fun GradleRunner.appendArguments(vararg arguments: String) =